feat: Add user model
This commit adds the user model in the `user.model.js` file. It defines the schema for the user collection in the MongoDB database, including fields for username, email, fullName, contactNo, address, language, earning, rating, avatar, password, role, and refreshToken. It also includes pre-save and methods for password validation and token generation.

```