feat: Add isWorker middleware
This commit adds the `isWorker.middleware.js` file, which contains the middleware function `isWorker`. This function is responsible for checking if the authenticated user has the role of "worker". If the user is not a worker, an error is thrown. The purpose of this commit is to add a middleware to protect routes that require the user to be a worker.