feat: Add isCustomer middleware
This commit adds the `isCustomer.middleware.js` file, which contains the middleware function `isCustomer`. This function is responsible for checking if the authenticated user has the role of "customer". If the user is not a customer, an error is thrown. The purpose of this commit is to add a middleware to protect routes that require the user to be a customer.