feat: Add itemDescription and qty fields to createPickup function
This commit adds the `itemDescription` and `qty` fields to the `createPickup` function in the `pickup.controller.js` file. These fields are now included in the request body and used to create a new pickup record in the database. This change enables the storage of additional information about the items being picked up.