feat: Accept request and send email notifications
This commit adds the functionality to accept a request in the `AccptRequest` function of the `request.controller.js` file. It updates the status of the request to "accepted" and sends email notifications to the user and the worker involved in the request. This change enables the successful acceptance of requests and improves communication between users and workers.