feat: Add @stripe/stripe-js dependency and makePayment function
This commit adds the "@stripe/stripe-js" dependency to the package.json and package-lock.json files, with a version of "^4.1.0". It also adds the "makePayment" function to the pickup.controller.js file, which handles the initiation of payment for a pickup. This change enables the integration of Stripe for payment processing in the application.