feat: Add owner to pickupDetails response
This commit adds the owner field to the response of the pickupDetails endpoint in the pickup.controller.js file. The owner information is retrieved from the User model and included in the ApiResponse object. This change enables users to retrieve the owner information along with the pickup and worker details.