feat: Refactor viewRequests controller to use aggregation pipeline
This commit refactors the viewRequests controller in the request.controller.js file to use the aggregation pipeline instead of multiple query stages. By using the aggregation pipeline, the code becomes more efficient and easier to read. This change improves the performance of fetching requests and ensures that only pending requests associated with the specified pickup are returned in the response.